+++ pymongo/synchronous/cursor_base.py
@@ -21,7 +21,7 @@ from collections.abc import Mapping, Sequence
 from typing import TYPE_CHECKING, Any, Optional, Union
 
 from pymongo import _csot
-from pymongo.cursor_shared import _AgnosticCursorBase
+from pymongo.cursor_shared import _CURSOR_DOC_FIELDS, _AgnosticCursorBase, _split_message
 from pymongo.lock import _create_lock
 from pymongo.message import _GetMore, _OpMsg, _Query
 from pymongo.response import PinnedResponse, Response
@@ -35,22 +35,6 @@ if TYPE_CHECKING:
     from pymongo.synchronous.pool import Connection
 
 _IS_SYNC = True
-
-_CURSOR_DOC_FIELDS = {"cursor": {"firstBatch": 1, "nextBatch": 1}}
-
-
-def _split_message(
-    message: Union[tuple[int, bytes], tuple[int, bytes, int]],
-) -> tuple[int, bytes, int]:
-    """Return request_id, data, max_doc_size.
-
-    :param message: (request_id, data, max_doc_size) or (request_id, data)
-    """
-    if len(message) == 3:
-        return message  # type: ignore[return-value]
-    # get_more and kill_cursors messages don't include BSON documents.
-    request_id, data = message  # type: ignore[misc]
-    return request_id, data, 0
 
 
 def _operation_to_command(
@@ -119,7 +103,12 @@ class _CursorBase(_AgnosticCursorBase[_DocumentType]):
         operation: Union[_Query, _GetMore],
         read_preference: _ServerMode,
     ) -> Response:
-        """Execute a cursor operation on the given connection and return a Response."""
+        """Execute a cursor operation on the given connection and return a Response.
+
+        :param conn: A Connection instance.
+        :param operation: A _Query or _GetMore object.
+        :param read_preference: The read preference to use.
+        """
         client = self._collection.database.client
         use_cmd = operation.use_command(conn)
         more_to_come = bool(operation.conn_mgr and operation.conn_mgr.more_to_come)
